requirements.txt
pyspark>=3.5.0
psycopg2-binary>=2.9.9
pandas>=2.2.2
pyyaml>=6.0.2
streamlit>=1.36.0
